Add CreateBet method

+++ Data/Bet.cs
@@ -13,11 +13,15 @@ public partial class Bet
     public Guid? IdEvent { get; set; }
 
     public DateTime? CreateDateBet { get; set; }
-
+//  0 - 1я команда победит
+//  1 - ничья
+//  2 - 2я команда победит
     public int? CoeffType { get; set; }
 
-    public int? Coeff { get; set; }
+//  Коэффициент на момент создания ставки
+    public double? Coeff { get; set; }
 
+//  Количество поставленных денег
     public double? BetAmount { get; set; }
 
     public virtual SportsEvent? IdEventNavigation { get; set; }
